+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.2 (2025-03-23)
 
+- feat: Added Homebrew package support (`brew tap kolarski/s-screen && brew install s-screen`)
 - feat: Added Arch Linux AUR package support (`yay -S s-screen`)
 - docs: Added website documentation link to package